Разбивает на js модули
